refactor: move requests logic to controllers
