+++ src/components/Header/Header.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { v4 as uuidv4 } from "uuid";
-import requestInstance from "../../requestInstance/requestInstance";
+import getCard from "../../controllers/getCard";
 import { AppBar, Toolbar, Typography, Button } from "@mui/material";
 import ViewComfyIcon from "@mui/icons-material/ViewComfy";
 
@@ -8,8 +8,7 @@ const iconStyles = { mr: 2 };
 
 const Header = ({ cardList, setCardList }) => {
   const handleAdd = async () => {
-    const randNum = Math.ceil(Math.random() * 100);
-    const { data: card } = await requestInstance.get(`/posts/${randNum}`);
+    const card = await getCard();
     const { title, body: text } = card;
     setCardList([...cardList, { id: uuidv4(), title, text }]);
   };
