+++ src/controllers/postCard.js
@@ -1,0 +1,17 @@
+import requestInstance from "../requestInstance/requestInstance";
+
+const postCard = async (card) => {
+  requestInstance
+    .post("/posts", {
+      userId: 1,
+      id: 101,
+      title: card.title,
+      body: card.text,
+    })
+    .then((res) => console.log("Request status: ", res.status))
+    .catch((e) => {
+      throw new Error(e.message);
+    });
+};
+
+export default postCard;
